Triangle metric corrected

+++ 2_NewDistance/shaders/quadtree_render_wireframe.glsl
@@ -23,7 +23,7 @@ layout (binding = LEAF_IDX_B) uniform idx_block {
 
 #ifndef LOD_GLSL
 uniform int morph;
-uniform float cpu_lod;
+uniform int cpu_lod;
 uniform int poly_type;
 #endif
